Cast mousePos to int before clamping.

+++ pixie_osx.mm
@@ -327,8 +327,8 @@ bool Window::PlatformUpdate()
     // Update mouse cursor position.
     NSPoint mousePos;
     mousePos = [window mouseLocationOutsideOfEventStream];
-    m_mouseX = std::clamp(mousePos.x, 0, m_width);
-    m_mouseY = std::clamp(m_height - mousePos.y - 1, 0, m_height);
+    m_mouseX = std::clamp((int)mousePos.x, 0, m_width);
+    m_mouseY = std::clamp(m_height - (int)mousePos.y - 1, 0, m_height);
 
     m_mouseX /= m_scalex;
     m_mouseY /= m_scaley;
